Refactor TwoFactorAuthenticatorsController#destroy
